+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add normalization option to WPM method
 - XYZMethod() now defines the default normalization for the possible copeland calls.
 - Test code coverage increased.
+- Added isfinite(), isnan(), one() for GreyNumber.
 
 ### 0.7.7